Perform the following actions
When a result is displayed, pressing a new digit clears the result and
starts a new calculation instead of appending the digit to the existing
result.

Pressing a number, an operator, a second number and a second operator
evaluates the initial pair of numbers and use that result as the first
number in a new calculation, along with the next operator.

When consecutive operators are entered, only the last operator is used
for the next operation.